% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.boot.FUN.mean.R
\name{create.boot.FUN.mean}
\alias{create.boot.FUN.mean}
\title{Bootstrap to get standard error of mean MESOR, amplitude and acrophase}
\usage{
create.boot.FUN.mean(contrast.frm, pairwise = TRUE)
}
\arguments{
\item{contrast.frm}{a string formula specifying the names of the predictors over which emmeans are desired.}

\item{pairwise}{If TRUE (default), test for differential rhythmicity (compare groups). If FALSE, test for rhythmicity only (detect rhythmicity)}
}
\value{
function input for [bootMer] calculating the mean of MESOR, Amplitude and Acrophase
}
\description{
Bootstrap function takes in a fitted merMod object as input and returns the statistics (e.g. standard error) of mean
MESOR, amplitude and acrophase in each group.
}
\examples{
\dontrun{
#' data(db.cosinor)
head(db.cosinor)
## create cosinor paramenters needed for cosinor model
db.model<-create.cosinor.param(time="Hour_of_Day", period=24, data=db.cosinor)

## examine the effects of gender on MESOR, acrophase, amplitude on hrv
f <- fit.cosinor.mixed(y = "hrv",
                       x = "gender",
                       random = "1|participant_id",
                       data = db.model)

boot.mean <- lme4::bootMer(f,
                           FUN = create.boot.FUN.mean(contrast.frm='~gender'),
                           nsim = 500,
                           parallel = "multicore",
                           ncpus = 8)
}
}
