% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.cosinor.param.R
\name{create.cosinor.param}
\alias{create.cosinor.param}
\title{Create cosinor parameters using the time variable and the cosinor period}
\usage{
create.cosinor.param(time, period, data)
}
\arguments{
\item{time}{a vector of time variable name, with same units as that of the cosinor period}

\item{period}{the length of the cosinor period, eg.24}

\item{data}{the dataframe with the time and period information}
}
\value{
\code{data.frame} with cosinor parameters. \code{rrr} = cosine component, \code{sss} = sine component
}
\description{
Create cosinor parameters using the time variable and the cosinor period
}
\examples{
data(db.cosinor)
head(db.cosinor)
## create cosinor paramenters needed for cosinor model
db.model<-create.cosinor.param(time="Hour_of_Day", period=24, data=db.cosinor)
}
