% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.means.ci.cosinor.R
\name{get.means.ci.cosinor}
\alias{get.means.ci.cosinor}
\title{Get means and confidence interval for MESOR, Amplitude, Acrophase}
\usage{
get.means.ci.cosinor(
  fit,
  contrast.frm,
  pairwise = FALSE,
  nsim = 500,
  parallel = "multicore",
  ncpus = 8,
  conftype = "norm",
  conflevel = 0.95,
  seed = NULL,
  ...
)
}
\arguments{
\item{fit}{the object from lmer() function}

\item{contrast.frm}{a string formula specifying the names of the predictors over which emmeans are desired.}

\item{pairwise}{If TRUE (default), test for differential rhythmicity (compare groups). If FALSE, test for rhythmicity only (detect rhythmicity)}

\item{nsim}{the number of simulations, positive integer; the bootstrap B (or R) for bootMer function.Default is 500.}

\item{parallel}{the type of parallel operation to be used (if any) for bootMer function. The default is "multicore".}

\item{ncpus}{integer: number of processes to be used in parallel operation: typically one would choose this to be the number of available CPUs.Default is 8.}

\item{conftype}{A character string representing the type of interval required.The value must be one of "norm", "basic","perc".}

\item{conflevel}{The confidence level required, default is 0.95.}

\item{seed}{argument to pass into 'seed' parameter in bootMer() function. Default NULL}

\item{...}{additional argument(s) for methods.}
}
\value{
Output with confidence intervals and means for MESOR, amplitude and acrophase
}
\description{
The function will takes in a fitted merMod object, transform the linear/raw coefficients to COSINOR non-linear/transformed coefficients
for MESOR, Amplitude and acrophase and estimated the standard error through bootstrapping to provide the
confidence interval.
}
\examples{
data(db.cosinor)
head(db.cosinor)
db.model <- create.cosinor.param(time = "Hour_of_Day", period = 24, data = db.cosinor)

# Rhythmicity stats (individual comparisons)
f1 <- lme4::lmer(hrv~rrr+sss+(1|participant_id),
               data=db.model, na.action = na.omit)

get.means.ci.cosinor(fit = f1, contrast.frm = '~ rrr + sss', nsim = 50, ncpus = 2, pairwise = TRUE)

# Differentiaal rhythmicity stats (pairwise comparisons)
f2 <- fit.cosinor.mixed(y = "hrv", x = "gender", random = "1|participant_id", data = db.model)
summary(f2)
get.means.ci.cosinor(fit = f2, contrast.frm = "~gender", nsim = 50, ncpus = 2, pairwise = FALSE)
get.means.ci.cosinor(fit = f2, contrast.frm = "~gender", nsim = 50, ncpus = 2, pairwise = TRUE)

}
