% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.cosinor.lmer.R
\name{ggplot.cosinor.lmer}
\alias{ggplot.cosinor.lmer}
\title{Plot a cosinor model}
\usage{
\method{ggplot}{cosinor.lmer}(
  object = NULL,
  x_str = NULL,
  period = NULL,
  db.means = NULL,
  DATA = NULL
)
}
\arguments{
\item{object}{a lmer fit object (lmerMod object)}

\item{x_str}{a character vector naming the covariate(s) to be plotted}

\item{period}{the length of time for a cycle.}

\item{db.means}{the data frame with mean values for MESOR, amplitude, and acrophase, for each comparison group}

\item{DATA}{a data frame containing the variables named in lmer object}
}
\value{
ggplot
}
\description{
Given a lmer model fit, generate a plot of the data with the fitted values.
}
\examples{
data(db.cosinor)
head(db.cosinor)
db.model <- create.cosinor.param(time = "Hour_of_Day", period = 24, data = db.cosinor)

f <- fit.cosinor.mixed(y = "hrv", x = "gender", random = "1|participant_id", data = db.model)
db.means <- get.means.ci.cosinor(fit = f, contrast.frm = "~gender", nsim = 50, ncpus = 2)
ggplot.cosinor.lmer(object=f,
                    x_str="gender",
                    period=24,
                    db.means=db.means,
                    DATA=db.model)

}
