% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/just.get.means.cosinor.R
\name{just.get.means.cosinor}
\alias{just.get.means.cosinor}
\title{Get MESOR(M), Amplitude(A), Acrophase}
\usage{
just.get.means.cosinor(fit, contrast.frm, pairwise = TRUE, ...)
}
\arguments{
\item{fit}{the lmer object from lmer()}

\item{contrast.frm}{a string formula specifying the names of the predictors over which emmeans are desired.}

\item{pairwise}{If TRUE (default), test for differential rhythmicity (compare groups). If FALSE, test for rhythmicity only (detect rhythmicity)}

\item{...}{other arguments passed on to methods}
}
\value{
MESOR, amplitude and acrophase estimates
}
\description{
Get the estimated marginal means for the specified factors or factor combinations in a linear model,
and transform to non-linear parameters MESOR, amplitude, and acrophase.
}
\examples{
require(lme4)
db.model<-create.cosinor.param(time="Hour_of_Day", period=24, data=db.cosinor)


# Rhythmicity stats (individual comparisons)
f1<-lme4::lmer(hrv~rrr+sss+(1|participant_id),
               data=db.model, na.action = na.omit)

just.get.means.cosinor(fit=f1, contrast.frm='~rrr+sss', pairwise = FALSE)

# Differential rhythmicity stats (pairwise comparisons)
f2<-lme4::lmer(hrv~gender+
               gender*rrr+
               gender*sss+(1|participant_id),
               data=db.model, na.action = na.omit)

just.get.means.cosinor(fit=f2, contrast.frm='~gender', pairwise = TRUE)
just.get.means.cosinor(fit=f2, contrast.frm='~gender', pairwise = FALSE)

}
