% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_rhythm.R
\name{simulate_rhythm}
\alias{simulate_rhythm}
\title{Simulate rhythmic data}
\usage{
simulate_rhythm(
  time_series = NULL,
  time_gap = 1,
  period = 24,
  amplitude = 0.5,
  acrophase = 12,
  mesor = 1,
  noise = NULL,
  plot = FALSE,
  ...
)
}
\arguments{
\item{time_series}{numeric vector for time (x-axis). If NULL (default), a [0, 24] hour time series will be created by `time_gap` steps}

\item{time_gap}{Difference between 2 time points. Used only when `time_series = NULL`. (default = 1)}

\item{period}{Period of rhythmic data (default = 24)}

\item{amplitude}{Amplitude of rhythmic data (default = 0.5)}

\item{acrophase}{Acrophase (peak time) of rhythmic data (default = 12)}

\item{mesor}{MESOR of rhythmic data (default = 1)}

\item{noise}{Add Gaussian noise to data (default = NULL). Passed to `sd` parameter of `rnorm`.}

\item{plot}{Boolean to plot of simulated data. (default = FALSE)}

\item{...}{Further parameters for `plot`}
}
\value{
Simulated rhythmic data (y-axis)
}
\description{
Simulate rhythmic data using cosinor regression.
}
\examples{

set.seed(666)
simulate_rhythm()
simulate_rhythm(plot = TRUE)
simulate_rhythm(time_gap = 0.5, plot = TRUE)
simulate_rhythm(time_series = 1:48, plot = TRUE)
simulate_rhythm(time_series = 1:24, period = 12, plot = TRUE)
simulate_rhythm(time_series = 1:24, amplitude = 1, plot = TRUE)
simulate_rhythm(time_series = 1:24, acrophase = 3, plot = TRUE)
simulate_rhythm(time_series = 1:24, mesor = 10, plot = TRUE)

par(mfrow = c(2,2))
simulate_rhythm(time_series = 1:24,
                mesor = 10,
                amplitude = 1,
                acrophase = 6,
                plot = TRUE,
                main = "noise = 0")
simulate_rhythm(time_series = 1:24,
                mesor = 10,
                amplitude = 1,
                acrophase = 6,
                plot = TRUE,
                noise = 0.1,
                main = "noise = 0.1")
simulate_rhythm(time_series = 1:24,
                mesor = 10,
                amplitude = 1,
                acrophase = 6,
                plot = TRUE,
                noise = 0.5,
                main = "noise = 0.5")
simulate_rhythm(time_series = 1:24,
                mesor = 10,
                amplitude = 1,
                acrophase = 6,
                plot = TRUE,
                noise = 1,
                main = "noise = 1.0")
}
